requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
colorama==0.3.9
feedparser==5.2.1
idna==2.6                 # via requests
listparser==0.18
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0               # via python-dateutil
urllib3==1.22             # via requests
